Moved test files around
